requirements.txt
torch==2.0.0
Pillow
numpy
scipy
torchvision==0.15.1
matplotlib
gdown